Wanted to remove the last functions from the Model,
by replacing the lazy tree with a non-lazy equivalent. The Tree.Zipper
seems like the ideal representation of a menu with one selected item,
but I'm stuck on how to render that, or transform it into another Tree.

Can't start at the root of the tree and recursively deal with the subtrees,
because then we've discarded the 'focus' of the zipper.

Can't(?) start at the focus and recurse back to the root... maybe? Could
be I'm just not seeing how to do that.

Writing this out just now, and skimming through <http://learnyouahaskell.com/zippers>,
it seems generating a "Breadcrumb trail" from the Zipper might be the
way to go?